+++ src/components/SearchInput.vue
@@ -1,0 +1,34 @@
+<script setup lang="ts">
+import { inject, Ref } from 'vue'
+const props = defineProps<{
+  label?: string
+}>()
+
+const label = props.label !== undefined ? props.label : 'Search'
+const searchTerm = inject<Ref<string>>('searchTerm')
+</script>
+<script lang="ts">
+export default {
+  inheritAttrs: false,
+}
+</script>
+<template>
+  <slot
+    v-bind="{
+      searchTerm,
+    }"
+  >
+    <label for="jsonsearchinput">
+      {{ label }}
+    </label>
+    <input
+      v-bind="$attrs"
+      name="jsonsearchinput"
+      class="jsonsearchinput"
+      autocomplete="off"
+      placeholder="Search"
+      type="text"
+      v-model="searchTerm"
+    />
+  </slot>
+</template>