+++ src/main.ts
@@ -1,6 +1,6 @@
 import { createApp } from 'vue'
 import App from './App.vue'
-import './assets/index.postcss'
+import '@/assets/index.postcss'
 
 const app = createApp(App)
 app.mount('#app')