+++ CHANGES.md
@@ -1,10 +1,21 @@
 # Changelog
 
+## 0.2.0 (2022-01-27)
+### New Features
+
+- Made all markup fully customizable through slots.
+- Added following named exports: `JsonSearch`, `ResultList`, `ResultListItem`, `ResultTitle`, `SearchInput`, `SearchResults`
+
+### Changed
+
+- Removed default import.
+- Changed default CSS to use classes instead of ids to allow multiple components on a page.
+
 ## 0.1.0 (2022-01-26)
 
-- Added configuration options (#2)
-- Added documentation (#1)
-- Added proper tests
+- Added configuration options. (#2)
+- Added documentation. (#1)
+- Added proper tests.
 ## 0.0.0 (2022-01-25)
 
-- Initial version
+- Initial version.